Refactor MCP large class responsibilities
- Split descriptor catalog validation into focused validators
- Split bootstrap tool specification tests by result and elicitation flows
- Extract sharding workflow input and lifecycle planning helpers
- Extract MCP E2E runtime, Docker, LLM readiness and trace helpers
- Split JDBC metadata loader tests by focused scenarios
